requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
python-jose
python-multipart
email-validator
python-dotenv
openai>=1.12.0
requests
google-auth
cloudinary
